\name{print.adk}
\alias{print.adk}
\title{ Print Function for Objects of Class adk }
\description{
 This print function is invoked for printing formatted output of class adk as it
is output from adk.test or adk.combined.test. 
}
\usage{
print.adk(x,...)
}
\arguments{
  \item{x}{ an object of class adk, as ouput by adk.test or adk.combined.test }
  \item{\dots}{further arguments passed to or from other methods}
}
\details{
The formatted output is different for objects generated by adk.test and
for object generated by adk.combined.test.
}


\author{Fritz Scholz}

\seealso{ \code{\link{adk.test}} 
and \code{\link{adk.combined.test}}}
\examples{
## Create input list of 3 sample vectors.
x <- list(c(1,3,2,5,7),c(2,8,1,6,9,4),c(12,5,7,9,11))
adk.test(x) 
}
\keyword{print}
\keyword{classes}

